Rename services > modules and nest in folder
