+++ balancer-js/src/modules/swaps/helpers.ts
@@ -2,9 +2,9 @@ import { BigNumberish, BigNumber } from '@ethersproject/bignumber';
 import { Zero, WeiPerEther } from '@ethersproject/constants';
 
 import { SwapType } from './types';
-import { isSameAddress } from '../utils';
+import { isSameAddress } from '@/utils';
 
-/*
+/**
  * Helper to create limits using a defined slippage amount.
  * @param tokensIn - Array of token in addresses.
  * @param tokensOut - Array of token out addresses.
