+++ balancer-js/tsconfig.json
@@ -9,7 +9,13 @@
         "esModuleInterop": true,
         "skipLibCheck": true /* Skip type checking of declaration files. */,
         "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */,
-        "resolveJsonModule": true
+        "resolveJsonModule": true,
+        "baseUrl": ".",
+        "paths": {
+          "@/*": [
+            "src/*"
+          ]
+        }
     },
     "include": ["./src", "src/abi/*.json"]
 }
